license: Add new untested hash condensor
To allow matching properly, we reduce an input file to a set of non-blank
lines, with no whitepsace in each line. This will allow more accurate hash
matching with licenses.spdx, as well as opening the door to pattern matching
each set of lines to work out the license from context.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>